Refactor file system providers to extend a new ReadOnlyFileSystemProvider
- Created a new abstract class ReadOnlyFileSystemProvider to encapsulate common read-only behavior for various file system providers.
- Updated CmisFileSystemProvider, HdfsFileSystemProvider, S3FileSystemProvider, SmbFileSystemProvider, and WebDavFileSystemProvider to extend ReadOnlyFileSystemProvider.
- Removed redundant methods and exceptions related to read-only operations from individual providers.
- Simplified the implementation of read attributes and byte channel creation in the new base class.
